refactor: wave15 peels — settings-scope, archive-planning, project-store-ops
Behavior-preserving package organization after wave 14:
- types/settings-scope.ts — global/project settings (~2.2k lines)
- types/archive-planning.ts — archive entries, mesh/planning, multi-project types
- rename remaining-ops-1 → task-store/project-store-ops.ts (last numbered ops gone)
- Ratchet types.ts line-count ceiling (~5872 → ~3074)
