+++ packages/core/src/__tests__/postgres/workflow-create.pg.test.ts
@@ -16,7 +16,7 @@ import type { AsyncDataLayer } from "../../postgres/data-layer.js";
 import type { TaskStore } from "../../store.js";
 import * as schema from "../../postgres/schema/index.js";
 import { writeProjectConfig } from "../../task-store/async-settings.js";
-import { __setWorkflowDefinitionBeforeInsertForTesting } from "../../task-store/remaining-ops-1.js";
+import { __setWorkflowDefinitionBeforeInsertForTesting } from "../../task-store/project-store-ops.js";
 
 import {
   pgDescribe,
