+++ scripts/line-count-baseline.json
@@ -17,7 +17,8 @@
   "packages/core/src/postgres/schema/project.ts": 2227,
   "packages/core/src/postgres/sqlite-migrator.ts": 2308,
   "packages/core/src/store.ts": 2820,
-  "packages/core/src/types.ts": 5872,
+  "packages/core/src/types.ts": 3074,
+  "packages/core/src/types/settings-scope.ts": 2287,
   "packages/dashboard/app/App.tsx": 2136,
   "packages/dashboard/app/components/AgentDetailView.tsx": 5567,
   "packages/dashboard/app/components/AgentsView.tsx": 2254,
@@ -93,11 +94,11 @@
   "packages/engine/src/__tests__/triage.test.ts": 7042,
   "packages/engine/src/agent-heartbeat.ts": 5422,
   "packages/engine/src/agent-tools.ts": 5437,
-  "packages/engine/src/executor.ts": 20411,
+  "packages/engine/src/executor.ts": 20489,
   "packages/engine/src/merger-ai.ts": 2445,
   "packages/engine/src/merger.ts": 11244,
   "packages/engine/src/pi.ts": 2900,
-  "packages/engine/src/project-engine.ts": 5527,
+  "packages/engine/src/project-engine.ts": 5580,
   "packages/engine/src/runtimes/in-process-runtime.ts": 2160,
   "packages/engine/src/scheduler.ts": 3666,
   "packages/engine/src/self-healing.ts": 12888,
